+++ src/os2datascanner/projects/report/reportapp/views/statistics_views.py
@@ -22,7 +22,7 @@ from dateutil.relativedelta import relativedelta
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.postgres.aggregates import StringAgg
 from django.core.exceptions import PermissionDenied
-from django.db.models import Q, Count, DateField, When, Case, CharField, Value
+from django.db.models import Q, Count, DateField, When, Case, CharField, Value, F
 from django.db.models.functions import Coalesce, TruncMonth
 from django.http import HttpResponseForbidden, Http404
 from django.utils.translation import gettext_lazy as _
@@ -145,9 +145,24 @@ class DPOStatisticsPageView(LoginRequiredMixin, TemplateView):
                 descendant_units = selected_unit.get_descendants(include_self=True)
                 positions = Position.employees.filter(unit__in=descendant_units)
                 accounts = Account.objects.filter(positions__in=positions).distinct()
-                self.matches = self.matches.filter(
-                    alias_relation__account__in=accounts).exclude(
-                    alias_relation__shared=True)
+
+                self.matches = (
+                    self.matches.annotate(
+                        total_relations=Count(
+                            'alias_relation',
+                            distinct=True
+                        ),
+                        shared_relations=Count(
+                            'alias_relation',
+                            filter=Q(alias_relation__shared=True),
+                            distinct=True
+                        )
+                    )
+                    # Ensure at least one relation is to these accounts
+                    .filter(alias_relation__account__in=accounts)
+                    # Keep only if not all relations are shared (i.e., at least one is not shared)
+                    .filter(~Q(total_relations=F('shared_relations')))
+                )
             else:
                 raise OrganizationalUnit.DoesNotExist(
                     _("An organizational unit with the UUID '{0}' was not found.".format(orgunit)))